repo: move sources to sep folder
