+++ src/bmp-conv.h
